+++ biblioteca/src/main/java/com/residencia/biblioteca/services/EditoraService.java
@@ -19,7 +19,7 @@ public class EditoraService {
 	}
 
 	public Editora buscarEditoraPorId(Integer id) {
-		return editoraRepo.findById(id).get();
+		return editoraRepo.findById(id).orElse(null);
 	}
 
 	public Editora salvarEditora(Editora editora) {
@@ -31,8 +31,24 @@ public class EditoraService {
 		return editoraRepo.save(editora);
 	}
 
-	public void deletarEditora(Editora editora) {
+	public Boolean deletarEditora(Editora editora) {
+	
+		if (editora == null)
+			return null;
+
+		Editora editoraExistente = buscarEditoraPorId(editora.getCodigoEditora());
+
+		if (editoraExistente == null)
+			return null;
+
 		editoraRepo.delete(editora);
+		
+		Editora editoraContinuaExistindo = buscarEditoraPorId(editora.getCodigoEditora());
+		 if (editoraContinuaExistindo == null)
+			 return true;
+		 
+		 else 
+			 return false;
 	}
 
 }
